+++ src/main/java/com/skycat/mystical/ConfigModel.java
@@ -40,6 +40,7 @@ public class ConfigModel {
     @Nest public BoldSlimesConfig boldSlimes = new BoldSlimesConfig();
     @Nest public ChangingArmorHurtsConfig changingArmorHurts = new ChangingArmorHurtsConfig();
     @Nest public SoundSwapConfig soundSwap = new SoundSwapConfig();
+    @Nest public FishingRodSwapConfig fishingRodSwap = new FishingRodSwapConfig();
 
     @SectionHeader("Logging") // Note: Logging as ERROR level does not always mean a critical error.
     public LogLevel failedToSetNightTimerLogLevel = LogLevel.WARN;
@@ -386,18 +387,28 @@ public class ConfigModel {
         public LogLevel logLevel = LogLevel.OFF;
         @PredicateConstraint("weightPredicate")
         public double weight = 1;
-        @PredicateConstraint("chancePredicate")
-        public double chance = 5.0;
         @PredicateConstraint("positiveNonzeroIntPredicate")
         public int numberOfSwaps = 50;
         public static boolean weightPredicate(double value) {
             return ConfigModel.weightPredicate(value);
         }
+        public static boolean positiveNonzeroIntPredicate(int value) {
+            return ConfigModel.positiveNonzeroIntPredicate(value);
+        }
+    }
+
+    public static class FishingRodSwapConfig {
+        public boolean enabled = true;
+        public LogLevel logLevel = LogLevel.OFF;
+        @PredicateConstraint("weightPredicate")
+        public double weight = 1;
+        @PredicateConstraint("chancePredicate")
+        public double chance = 100.0;
         public static boolean chancePredicate(double value) {
             return ConfigModel.chancePredicate(value);
         }
-        public static boolean positiveNonzeroIntPredicate(int value) {
-            return ConfigModel.positiveNonzeroIntPredicate(value);
+        public static boolean weightPredicate(double value) {
+            return ConfigModel.weightPredicate(value);
         }
     }