+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - A lot of internal changes
 - Haven cost is now configurable
 - Added a bunch of new translations
+- Removed unused "chance" config for soundSwap
+- Added fishingRodSwap